fix: move themes to the themes file
